Move the HTTP Headers management to \Nova\Net\Response
